Python/notify_Discord/requirements.txt
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
